fix(Constant): 당첨 금액 형변환(Integer -> Long)에 따른 코드 수정

+++ src/main/java/lotto/domain/user/User.java
@@ -25,11 +25,11 @@ public class User {
     }
   }
 
-  private Integer getLottoPrice() {
+  private Long getLottoPrice() {
     return LottoPrice.LOTTO_PRICE.getPrice();
   }
 
   public int getTicketCount(){
-    return money / getLottoPrice();
+    return (int) (money / getLottoPrice());
   }
 }
